detailErrorModify,searchbarConnect
