+++ src/components/FeedItem.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { deleteDoc, doc, updateDoc } from 'firebase/firestore';
 import { db } from '../firebase';
 import styled from 'styled-components';
-import Modal from './detail/Detail';
+import Modal from '../shared/detail';
 
 const FeedItem = ({ feed, setFeeds }) => {
   const [isModalOpen, setIsModalOpen] = useState(false);
@@ -36,11 +36,15 @@ const FeedItem = ({ feed, setFeeds }) => {
   };
 
   const openModal = () => {
-    setIsModalOpen(true);
+    if (!isModalOpen) {
+      setIsModalOpen(true);
+    }
   };
 
   const closeModal = () => {
-    setIsModalOpen(false);
+    if (isModalOpen) {
+      setIsModalOpen(false);
+    }
   };
 
   return (
