+++ src/pages/Home.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { openModalHandler, Modal } from '.detail';
+import { openModalHandler, Modal } from '../shared/detail';
+
 
 function Home() {
   return (
